requirements.txt
altgraph==0.17.4
bottle==0.13.3
macholib==1.16.3
packaging==25.0
proxy_tools==0.1.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyobjc-framework-Security==11.0
pyobjc-framework-WebKit==11.0
pywebview==5.4
setuptools==80.3.0
typing_extensions==4.13.2
websockets==15.0.1
